fix: sanitize the create-side name arms and the workspace lookup needle
Review follow-up: addWorkspace and newWindow are the create-side twins of
the rename arms fixed in the previous commit — workspace new --name,
session new --workspace-name --create-workspace, and window new --name
still carried an interior newline into {AGT_WORKSPACE_NAME}/{AGT_WINDOW_NAME}.

The lookup needle in workspace(named:) sanitizes too, and in that method
rather than ensureWorkspace: session.new --workspace-name without
--create-workspace calls workspace(named:) directly, so a needle-side-only
fix there would leave a raw needle against a sanitized stored name and read
back as a silent "workspace not found". Sanitizing before the blank gate
also keeps ensureWorkspace idempotent for control-char-only input (blank →
nil, not a fresh unmatchable workspace per call).

The sanitize tests move to AppStoreNameSanitizeTests.swift — the new arms
pushed AppStoreTests.swift over the 2000-line file budget.
